Main program code/server/requirements.txt
imageio
PILLOW
pygifsicle
matplotlib
stitching
numpy
mysql.connector.python
torchvision
transformers
pytorch_lightning
supervision
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121